+++ src/icemac/ab/calexport/browser/renderer.py
@@ -1,5 +1,10 @@
 import grokcore.component as grok
 import icemac.ab.calendar.browser.renderer.table
+import icemac.ab.calendar.interfaces
+import icemac.ab.calexport.interfaces
+import icemac.addressbook.entities
+import zope.cachedescriptors.property
+import zope.security.proxy
 
 
 class ExportTable(icemac.ab.calendar.browser.renderer.table.Table):
@@ -15,8 +20,33 @@ class ExportEvent(icemac.ab.calendar.browser.renderer.table.TableEvent):
 
     default_text = u''  # Do not render events without title
 
+    @property
+    def masterdata(self):
+        # Without the following line we get a ForbiddenError for `__getitem__`
+        # when accessing the annotations where `IExportMasterdata` are stored.
+        # As only authorized users are able to access this function, this is no
+        # security hole.
+        unsave_calendar = zope.security.proxy.getObject(
+            icemac.ab.calendar.interfaces.ICalendar(self.context))
+        return icemac.ab.calexport.interfaces.IExportMasterdata(
+            unsave_calendar)
+
     def action_url(self):
         return None
 
     def info(self):
         return []
+
+    def dd_class(self):
+        special_field = self.masterdata.special_field
+        if not special_field:
+            return None
+
+        event = icemac.ab.calendar.interfaces.IEvent(self.context)
+        field = icemac.addressbook.entities.get_bound_schema_field(
+            event, None, special_field, default_attrib_fallback=False)
+        try:
+            is_special = field.get(field.context)
+        except AttributeError:  # Recurring event has no matching field:
+            is_special = False
+        return 'special' if is_special else None